update IPuzzle.Resolve() to return long for puzzles of day 3

+++ AdventOfCode/01/Puzzle1.cs
@@ -12,7 +12,7 @@ namespace AdventOfCode._01
             Target = target;
         }
 
-        public int Resolve(List<string> input)
+        public long Resolve(List<string> input)
         {
             var parsedInput = input.Select(int.Parse).ToList();
             parsedInput.Sort();
